% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.speciesList}
\alias{as.speciesList}
\alias{as.speciesList.data.frame}
\alias{as.speciesList.default}
\alias{as.speciesList.matrix}
\title{Coerce to species list}
\usage{
as.speciesList(x, ...)

\method{as.speciesList}{default}(x, ...)

\method{as.speciesList}{matrix}(x, ...)

\method{as.speciesList}{data.frame}(x, siteCol, speciesCol, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional parameters to various methods (e.g. for
class \code{data.frame})}

\item{siteCol}{Name of column indicating sites}

\item{speciesCol}{Name of column indicating species}
}
\value{
a \code{speciesList} object
}
\description{
Coerce to species list
}

