% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calcDimIds}
\alias{calcDimIds}
\title{Estimate and set dimnames names}
\usage{
calcDimIds(x, nDims)
}
\arguments{
\item{x}{A \code{\link{list}} object}

\item{nDims}{Number of dimensions into which variables are to be
clustered}
}
\value{
\code{x} with \code{\link{dimIds}}
}
\description{
Uses \code{\link{kmeans}} clustering on the first axis of a
correspondence analysis (\code{\link{corresp}}) of the union of
all dimnames against variables
}

