% Generated by roxygen2 (4.0.0): do not edit by hand
\name{combineDists}
\alias{combineDists}
\title{Euclidean combination of distance matrices}
\usage{
combineDists(dist1, dist2, a, p = 2, reorder = TRUE)
}
\arguments{
\item{dist1}{A distance matrix}

\item{dist2}{A second distance matrix}

\item{a}{Weighting parameter (number between 0 and 1 giving the
weighting towards \code{dist1})}

\item{p}{Metric exponent (\code{p = 2} for Euclidean)}

\item{reorder}{Passed to \code{\link{subscript.longDist}}}
}
\value{
combined distance matrix
}
\description{
Euclidean combination of distance matrices
}

