% Generated by roxygen2 (4.0.0): do not edit by hand
\name{conversion}
\alias{conversion}
\alias{conversion.character}
\alias{conversion.default}
\alias{conversion.logical}
\alias{conversion.matrix}
\title{Convert non-integer subscript to integer}
\usage{
conversion(i, nms, ...)

\method{conversion}{default}(i, nms, ...)

\method{conversion}{character}(i, nms, ...)

\method{conversion}{logical}(i, nms, ...)

\method{conversion}{matrix}(i, nms, ...)
}
\arguments{
\item{i}{non-integer subscript (or list of non-integer
subscripts for each dimension}

\item{...}{additional info about the object}

\item{nms}{list of dNames}
}
\value{
Subscripts in integer form
}
\description{
Convert non-integer subscript to integer
}

