% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dimIds}
\alias{dimIds}
\alias{dimIds<-}
\alias{setDimIds}
\title{Set and get dimnames names}
\usage{
dimIds(x)

dimIds(x) <- value

setDimIds(x, value)
}
\arguments{
\item{x}{An object}

\item{value}{Identifiers for object dimensions}
}
\description{
Set and get dimnames names
}

