% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dist.data.frame}
\alias{dist.data.frame}
\title{Distance matrix in long form}
\usage{
dist.data.frame(x, ...)
}
\arguments{
\item{x}{\code{\link{dist}} object}

\item{...}{not used}
}
\value{
\code{dist.data.frame} object.  A data frame with three
columns, \code{row},\code{col}, and \code{dist}
}
\description{
Distance matrix in long form
}

