% Generated by roxygen2 (4.0.0): do not edit by hand
\name{longDist}
\alias{as.longDist}
\alias{longDist}
\title{Distance matrix in long form}
\usage{
longDist(x, ...)

as.longDist(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{not used}
}
\value{
\code{dist.data.frame} object.  A data frame with three
columns, \code{row},\code{col}, and \code{dist}
}
\description{
Distance matrix in long form
}

