% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.dbDiversityRegression}
\alias{plot.dbDiversityRegression}
\title{Plot distance-based diversity regression}
\usage{
\method{plot}{dbDiversityRegression}(x, y, ...)
}
\arguments{
\item{x}{\code{\link{dbDiversityRegression}} object}

\item{y}{not used}

\item{...}{not used}
}
\description{
Plot distance-based diversity regression
}

