% Generated by roxygen2 (4.0.0): do not edit by hand
\name{poly.data.frame}
\alias{as.poly.data.frame}
\alias{as.poly.data.frame.default}
\alias{is.poly.data.frame}
\alias{poly.data.frame}
\title{Poly data frame}
\usage{
poly.data.frame(...)

is.poly.data.frame(x)

as.poly.data.frame(x, nDims)

\method{as.poly.data.frame}{default}(x, nDims)
}
\arguments{
\item{...}{Objects to combine into a \code{poly.data.frame}
object.}

\item{x}{An \code{R} object}

\item{nDims}{see \code{\link{calcDimIds}}}
}
\description{
Data frame with unusual shape
}

