% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.poly.data.frame}
\alias{print.poly.data.frame}
\title{Print poly data frame}
\usage{
\method{print}{poly.data.frame}(x, full = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link{poly.data.frame}} object.}

\item{full}{Should entire object be printed?}

\item{...}{Not used}
}
\description{
Print poly data frame
}

