% Generated by roxygen2 (4.0.0): do not edit by hand
\name{processSubscript}
\alias{processSubscript}
\title{Process subscript}
\usage{
processSubscript(x, i, ...)
}
\arguments{
\item{x}{an object to be subscripted}

\item{i}{subscript}

\item{...}{not used}
}
\value{
the processed subscript with a \code{processed} attribute
}
\description{
Process subscript
}

