% Generated by roxygen2 (4.0.0): do not edit by hand
\name{reorder.longDist}
\alias{reorder.longDist}
\title{Reorder distances as in a dist object}
\usage{
\method{reorder}{longDist}(x, X, ...)
}
\arguments{
\item{x}{\code{longDist} object}

\item{X}{character vector giving order}

\item{...}{not used}
}
\value{
vector reordered version of \code{x}
}
\description{
Reorder distances as in a dist object
}

