% Generated by roxygen2 (4.0.0): do not edit by hand
\name{reorder.poly.data.frame}
\alias{reorder.poly.data.frame}
\title{Reorder poly data frame}
\usage{
\method{reorder}{poly.data.frame}(x, X, ...)
}
\arguments{
\item{x}{a \code{\link{poly.data.frame}}}

\item{X}{optional \code{\link{dNames}} on which to reorder
\code{x}}

\item{...}{not used}
}
\value{
reordered \code{\link{poly.data.frame}}
}
\description{
Reorder poly data frame
}

