% Generated by roxygen2 (4.0.0): do not edit by hand
\name{speciesList}
\alias{speciesList}
\title{Species list}
\usage{
speciesList(x, nms, ...)
}
\arguments{
\item{x}{List of character vectors giving the names of species
present at various sites.}

\item{nms}{Optional names for the sites.}

\item{...}{Not used.}
}
\value{
\code{speciesList} object
}
\description{
Species list
}

