\name{subscript.data.frame}
\alias{subscript.data.frame}
\title{Subscript a data frame}
\usage{
\method{subscript}{data.frame}(x, i, ...)
}
\arguments{
  \item{x}{A data frame.}

  \item{i}{Indices for the rows.}

  \item{...}{Not used.}
}
\value{
A subscripted data frame
}
\description{
Subscript a data frame by rows (i.e. replicates) only.
}

