% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subscript.dist.data.frame}
\alias{subscript.dist.data.frame}
\title{Subscript a distance data frame}
\usage{
\method{subscript}{dist.data.frame}(x, i, ...)
}
\arguments{
\item{x}{\code{dist.data.frame} object}

\item{i}{subscript list}

\item{...}{Not used}
}
\value{
subscripted \code{dist.data.frame} object
}
\description{
Subscript a distance data frame
}

