\name{subscript.dist}
\alias{subscript.dist}
\title{Subscript a distance matrix}
\usage{
  \method{subscript}{dist} (x, i, ...)
}
\arguments{
  \item{x}{A \code{\link{dist}} object.}

  \item{i}{Indices for the objects.}

  \item{...}{Not used.}
}
\value{
  A subscripted \code{\link{dist}} object.
}
\description{
  Subscript a distance matrix as a single-dimensional
  object.
}

