% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subscript.longDist}
\alias{subscript.longDist}
\title{Subscript a distance data frame}
\usage{
\method{subscript}{longDist}(x, i, ...)
}
\arguments{
\item{x}{\code{longDist} object}

\item{i}{subscript list}

\item{...}{Not used}
}
\value{
subscripted \code{longDist} object
}
\description{
Subscript a distance data frame
}

