\name{subscript.poly.data.frame}
\alias{subscript.poly.data.frame}
\title{Subscript a poly data frame}
\usage{
\method{subscript}{poly.data.frame}(x, i, ...)
}
\arguments{
  \item{x}{\code{poly.data.frame} object}

  \item{i}{subscript list}
}
\value{
subscripted \code{poly.data.frame}
}
\description{
Subscript a poly data frame
}

