\name{subscript.speciesList}
\alias{subscript.speciesList}
\title{Subscript a species list}
\usage{
\method{subscript}{speciesList}(x, i, ...)
}
\arguments{
  \item{x}{\code{speciesList} object}

  \item{i}{Subscript list}
}
\value{
subscripted \code{specieslist}
}
\description{
Subscript a species list
}

