\name{dNamesNested}
\alias{dNamesConcat}
\alias{dNamesNested}
\alias{dimIdsConcat}
\alias{dimIdsNested}
\alias{dimIdsUnique}
\title{Utility functions for poly.data.frame}
\usage{
dNamesNested(x)

dNamesConcat(x)

dimIdsNested(x)

dimIdsConcat(x)

dimIdsUnique(x)
}
\arguments{
  \item{x}{A \code{\link{poly.data.frame}} object}
}
\description{
Utility functions for poly.data.frame
}

