% Generated by roxygen2 (4.0.0): do not edit by hand
\name{within.poly.data.frame}
\alias{within.poly.data.frame}
\title{Within a poly data frame}
\usage{
\method{within}{poly.data.frame}(data, expr, ...)
}
\arguments{
\item{data}{\code{\link{poly.data.frame}} object}

\item{expr}{an expression to evaluate within \code{data}}

\item{...}{passed to \code{\link{within}}}
}
\value{
\code{\link{poly.data.frame}} object
}
\description{
Within a poly data frame
}

