% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compareNames}
\alias{compareNames}
\title{Compare names}
\usage{
compareNames(nms1, nms2)
}
\arguments{
\item{nms1}{list of names vectors}

\item{nms2}{list of names vectors}
}
\value{
A \code{\link{matrix}} with rows representing list
elements of \code{nms1} and columns representing list elements of
\code{nms2}.  Entry \code{i, j} gives the number of elements that
are in \code{nms2[[j]]} but not in \code{nms1[[i]]} (in a
\code{setdiff(length(.))} way)
}
\description{
Compare names
}
\seealso{
\code{\link{setdiff}}, \code{\link{length}}
}

