% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dbDiversityProfile}
\alias{aHpd}
\alias{coef.dbDiversityRegression}
\alias{dbDiversityProfile}
\alias{dbDiversityRegression}
\alias{logLik.dbDiversityRegression}
\alias{mean.dbDiversityRegression}
\alias{posterior}
\title{Simple linear regression on distance-based diversity indices}
\usage{
dbDiversityProfile(slist, sdist1, sdist2, aGrid, reOrder = TRUE)

dbDiversityRegression(slist, sdist1, sdist2, resp, aGrid, reOrder = TRUE)

\method{logLik}{dbDiversityRegression}(object, ...)

posterior(object, ...)

\method{mean}{dbDiversityRegression}(x, ...)

\method{coef}{dbDiversityRegression}(object, ...)

aHpd(object, level = 0.95)
}
\arguments{
\item{slist}{A \code{\link{speciesList}} object}

\item{sdist1}{A distance matrix}

\item{sdist2}{A second distance matrix}

\item{resp}{A response variable}

\item{aGrid}{Optional grid of weighting parameters (see
\code{\link{combineDists}})}

\item{reOrder}{Passed to \code{\link{subscript.longDist}}}

\item{object}{A \code{\link{dbDiversityRegression}} object}

\item{...}{not used}

\item{x}{\code{\link{dbDiversityRegression}} object}

\item{level}{Size of the highest posterior density region.}
}
\value{
TODO

TODO

A data frame with two columns:  the values of the
	grid within the hpd region and the value of the posterior
	at each point in this grid.
}
\description{
Simple linear regression on distance-based diversity indices

Find points on a grid within a 100\code{p}\% highest posterior
density region for the tuning parameter, a
}
\details{
The 100\code{p}\% highest posterior density region for 'a' is
the subset of the interval between 0 and 1, which contains
100\code{p}\% of the probability.
}

