% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dimIds}
\alias{dimIds}
\alias{dimIds<-}
\alias{setDimIds}
\title{Set and get dimension identifiers}
\usage{
dimIds(x)

dimIds(x) <- value

setDimIds(x, value)
}
\arguments{
\item{x}{An object}

\item{value}{Identifiers for object dimensions}
}
\description{
Dimension identifiers (\code{dimIds}) allow combining objects into
\code{\link{poly.data.frame}} objects, which are related along the
dimensions of each of their consitituent objects.  In order to
make use of these relationships, the dimensions of each object
must be identified, which is accomplished by \code{dimIds}.
}
\details{
\code{dimIds} are stored in two different ways:
\itemize{
\item For \code{data.frame} objects and everything else
without \code{dimnames}, \code{dimIds} are stored as an attribute.
\item For any object with \code{dimnames} except \code{data.frame}s,
\code{dimIds} are stored as the names of the list of \code{dimnames}.
}
}

