\name{dimIds<-}
\alias{dimIds}
\alias{dimIds<-}
\alias{setDimIds}
\title{Set dimnames names}
\usage{
dimIds(x) <- value

setDimIds(x, value)

dimIds(x)
}
\arguments{
  \item{x}{An object}

  \item{value}{Identifiers for object dimensions}

  \item{x}{An object}
}
\value{
The object with dimension identifiers

Names of the dimnames of \code{x}
}
\description{
Set dimnames names

Get dimnames names
}

