\name{follows}
\alias{\%f\%}
\alias{follows}
\title{Follows}
\usage{
follows(f, g)

f \%f\% g
}
\arguments{
  \item{f}{a function of one argument}

  \item{g}{another function with range equal to the domain
  of \code{f}}

  \item{...}{additional arguments to \code{g}}
}
\value{
The function given by \code{f} follows \code{g} of
\code{x}.
}
\description{
Syntactic sugar for composition of two functions. Useful
for apply functions.
}
\examples{
sqr <- function(x) x^2
min10 <- function(x) x - 10
(sqr \%f\% min10)(2)
div5 <- function(x) x/5
(sqr \%f\% min10 \%f\% div5)(2)
(unique \%f\% unlist \%f\% `[`)(letters[1:10], c(3, 2))
}

