% Generated by roxygen2 (4.0.0): do not edit by hand
\name{longDist}
\alias{as.longDist}
\alias{as.longDist.dist}
\alias{longDist}
\title{Distance matrix in long form}
\usage{
longDist(x, ...)

as.longDist(x, ...)

\method{as.longDist}{dist}(x, norm = FALSE, reorder = FALSE,
  sortorder = FALSE, ...)
}
\arguments{
\item{x}{an object}

\item{...}{not used}

\item{norm}{normalize distances such that the maximum distance
equals one?}

\item{reorder}{see \code{\link{reorder.longDist}}}

\item{sortorder}{similar to \code{\link{reorder}} but with sorted
\code{\link{dNames}}}
}
\value{
\code{dist.data.frame} object.  A data frame with three
columns, \code{row},\code{col}, and \code{dist}
}
\description{
Distance matrix in long form
}

