% Generated by roxygen2 (4.0.0): do not edit by hand
\name{speciesList}
\alias{speciesList}
\title{Species list}
\usage{
speciesList(x, nms, ...)
}
\arguments{
\item{x}{List of character vectors giving the names of species
present at various sites.}

\item{nms}{Optional names for the sites.}

\item{...}{Not used.}
}
\value{
\code{speciesList} object.  A list with one element for
each site giving a character vector of the names of the species
present at that site.  Also includes an attribute
\code{"totalSpeciesList"} with the list of all species over all
sites.
}
\description{
Species list
}

