\name{subscript}
\alias{ss}
\alias{subscript}
\alias{subscript.default}
\title{Subscript}
\usage{
subscript(x, i, ...)

ss(x, i, ...)

\method{subscript}{default}(x, i)
}
\arguments{
  \item{x}{An object with data to be subscripted.}

  \item{...}{Indices for the dimensions of replication of
  \code{x}.}
}
\value{
A subscripted version of \code{x}.
}
\description{
Generic function for subscripting data objects.
}

