% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subscript.speciesList}
\alias{subscript.speciesList}
\title{Subscript a species list}
\usage{
\method{subscript}{speciesList}(x, i, ...)
}
\arguments{
\item{x}{\code{speciesList} object}

\item{i}{Subscript list}

\item{...}{Not used}
}
\value{
subscripted \code{specieslist}
}
\description{
Subscript a species list
}

