% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dNamesNested}
\alias{concat2NestedIndex}
\alias{dNamesConcat}
\alias{dNamesNested}
\alias{dimIdsConcat}
\alias{dimIdsNested}
\alias{dimIdsUnique}
\title{Utility functions for poly.data.frame}
\usage{
dNamesNested(x)

dNamesConcat(x)

dimIdsNested(x)

dimIdsConcat(x)

dimIdsUnique(x)

concat2NestedIndex(x)
}
\arguments{
\item{x}{A \code{\link{poly.data.frame}} object}
}
\description{
Utility functions for poly.data.frame
}

