\name{dNamesNested}
\alias{dIdsConcat}
\alias{dIdsNested}
\alias{dIdsUnique}
\alias{dNamesConcat}
\alias{dNamesNested}
\title{Utility functions for poly.data.frame}
\usage{
dNamesNested(x)

dNamesConcat(x)

dIdsNested(x)

dIdsConcat(x)

dIdsUnique(x)
}
\arguments{
  \item{x}{A \code{\link{poly.data.frame}} object}
}
\description{
Utility functions for poly.data.frame
}

